+++ src/components/Footer.tsx
@@ -20,6 +20,14 @@ const FooterDescription = () => (
     >
       Voir le code source
     </a>
+    . Illustrations de{" "}
+    <a
+      href="https://storyset.com/work"
+      target="_blank"
+      rel="noreferrer nofollow noopener"
+    >
+      Freepik Storyset
+    </a>
     .
   </>
 );
@@ -59,6 +67,10 @@ const footerBottomLinks = [
     link: "/mentions-legales",
     title: "Mentions légales",
   },
+  {
+    link: "/contact",
+    title: "Nous contacter",
+  },
 ];
 
 const Footer = () => (
@@ -66,7 +78,7 @@ const Footer = () => (
     <FooterBody description={<FooterDescription />}>
       <Logo splitCharacter={10}>République Française</Logo>
       <FooterOperator>
-        <img src="/images/cpam.png" alt="CNAM" width="200" height="66" />
+        <img src="/images/cnam.png" alt="CNAM" width="200" height="66" />
       </FooterOperator>
       {footerBodyLinks.map((item) => (
         <FooterBodyItem key={item.title}>