+++ src/services/faq/doctor/exchange.ts
@@ -1,0 +1,19 @@
+/* eslint-disable max-len */
+export default [
+  {
+    answer:
+      "<p>Pour permettre une prise en charge coordonnée, vous transmettez au patient, avec le courrier d’adressage, un courrier d'accompagnement destiné au psychologue (modèle disponible <a href='/documents/Courrier_d_adressage.docx' target='_blank'>ici</a>).  Ce document, que le psychologue conserve, a vocation à l’informer du contexte, des éléments cliniques et des motifs de l’adressage, éventuellement des scores ou d’échelles.</p><p>En retour, le psychologue vous informe du nombre de séances prévues avec le patient. </p>",
+    question:
+      "Quels échanges en début de prise en charge avec le psychologue ?",
+  },
+  {
+    answer:
+      "<p>Pour permettre un parcours de soins coordonné et adapté aux besoins du patient, le psychologue vous transmet le compte-rendu de fin de prise en charge, en accord avec le patient (ou le titulaire de l’autorité parentale). Vos coordonnées sont fournies par le patient.</p><p>Ce document sera, le cas échéant, le support d’une concertation adaptant la prise en charge.</p><p>Vous pouvez à tout moment échanger avec le psychologue qui accompagne votre patient, si vous le jugez nécessaire. Le psychologue est membre de l’équipe de soins au sens des articles L. 1110-4 et L. 1110-12 du code de la santé publique. À ce titre, il peut, en l’absence d’opposition du patient (ou du titulaire de l’autorité parentale), partager avec vous les informations strictement nécessaires à la coordination des soins.</p>",
+    question: "Un compte-rendu de fin de prise en charge m’est-il envoyé ?",
+  },
+  {
+    answer:
+      "<p>A la fin de l’accompagnement, vous vous concertez avec le psychologue partenaire pour réévaluer et adapter la prise en charge du patient. Vous associez à cette concertation un psychiatre en cas de non-amélioration de l’état initial après la réalisation des huit séances prises en charge, d’aggravation de l’état initial ou de doute sur le diagnostic.<p>Différents cas peuvent se présenter :<ul><li>la nécessité d’une consultation avec un psychiatre (en libéral ou à l’hôpital) pour une évaluation plus approfondie</li><li>la nécessité d’une prise en charge la plus adaptée: centre médico-psychologique (CMP), CMP infanto-juvéniles, service de psychiatrie ou pédopsychiatrie, maison des adolescents…</li><li>une évolution favorable ne nécessitant pas nécessairement de nouvel accompagnement psychologique</li><li>un nouvel accompagnement psychologique</li></ul></p><p>Dans ce dernier cas, vous pouvez ré-adresser le patient dans la limite de 8 séances remboursées par année civile. Vous vérifiez donc avec lui le nombre de séances réalisées dans l’année.<br/>Le patient peut aussi décider de poursuivre avec son psychologue dans un cadre non remboursé. Si celui-ci n’en a pas les moyens, vous pouvez l’orienter vers une structure hospitalière de type centre médico-psychologique.</p>",
+    question: "Quelle suite donner à la fin de la prise en charge ?",
+  },
+];