+++ src/services/faq/doctor/practices.ts
@@ -1,0 +1,15 @@
+/* eslint-disable max-len */
+export default [
+  {
+    answer:
+      "<p>Non, l’utilisation d’échelles et de scores n’est pas obligatoire.</p><p>Cette utilisation est recommandée notamment en cas de doute sur l’intensité de la souffrance psychique de votre patient.</p><p>Elle vous permet par ailleurs d’échanger avec le psychologue sur des éléments partagés.</p>",
+    question:
+      "L’utilisation des échelles d’évaluation de l’intensité des symptômes anxieux ou dépressifs est-elle obligatoire ?",
+  },
+  {
+    answer:
+      "<p>Avant d’orienter un patient vers un accompagnement psychologique, vous devez réévaluer les traitements médicamenteux de votre patient et en particulier ses traitements psychotropes. <p>La Haute Autorité de Santé (HAS) met par exemple en ligne des outils d’aide à la déprescription des benzodiazépines et médicaments apparentés :<ul><li>Fiche de bon usage des médicaments : <a href='https://www.has-sante.fr/upload/docs/application/pdf/2018-07/fiche_bum_benzodiazepines_anxiete_cd_27062018.pdf' target='_blank'>Quelle place pour les benzodiazépines dans l’anxiété ?</a></li><li>Recommandations de bonne pratique : <a href='https://www.has-sante.fr/jcms/c_601509/modalites-d-arret-des-benzodiazepines-et-medicaments-apparentes-chez-le-patient-age' target='_blank'>Modalités d'arrêt des benzodiazépines et médicaments apparentés chez le patient âgé</a></li></ul></p><p><b>A savoir</b> : si votre patient est éligible, mais qu’il est dans une des situations suivantes : <ul><li>traitement par antidépresseurs depuis plus de 3 mois </li><li>traitement par hypnotiques ou benzodiazépines depuis plus d’un mois</li><li>traitement par un autre psychotrope</li><li>patients bipolaires ou borderline sous antiépileptiques </li></ul></p><p>Avant de l’orienter vers un accompagnement psychologique, vous sollicitez l’avis d’un psychiatre. Celui-ci formalise un avis quant à la prise en charge la plus adaptée pour votre patient et réévalue la pertinence des psychotropes prescrits.</p>",
+    question:
+      "L’orientation vers un accompagnement psychologique est-elle l’occasion de réévaluer la pertinence des traitements médicamenteux ?",
+  },
+];