+++ src/services/faq/patient/children.ts
@@ -1,0 +1,15 @@
+/* eslint-disable max-len */
+export default [
+  {
+    answer:
+      "<p>Les psychologues qui souhaitent accompagner des enfants et adolescents sont listés dans l’annuaire. Il suffit de filtrer les résultats de l’annuaire par public cible : adultes ou enfants/adolescents (à venir).</p>",
+    question:
+      "Comment choisir un psychologue qui souhaite suivre des enfants et adolescents ?",
+  },
+  {
+    answer:
+      "<p>L’accord des titulaires de l’autorité parentale (parents ou tuteurs) est indispensable, si le patient est mineur. C’est le médecin qui est en charge de recueillir cet accord avant le début de l’accompagnement psychologique.</p>",
+    question:
+      "L’accord des parents est-il nécessaire pour voir un psychologue ?",
+  },
+];