+++ src/services/faq/patient/eligibility.ts
@@ -1,0 +1,13 @@
+/* eslint-disable max-len */
+export default [
+  {
+    answer:
+      "<p>A partir d’avril 2022, l’ensemble des personnes âgées de 3 ans ou plus (enfants, adolescents et adultes), souffrant de troubles psychiques d’intensité légère à modérée, peuvent, après consultation d’un médecin, bénéficier de la prise en charge d’un accompagnement psychologique réalisé par un psychologue partenaire.<p>Vous êtes concerné(e) :<ul><li>Si vous êtes angoissé(e), anxieux(se) ou déprimé(e)</li><li>Si vous êtes en situation de mal-être </li><li>Si vous avez un problème de consommation de tabac, d’alcool ou de cannabis </li><li>Si vous avez un trouble du comportement alimentaire </li></ul></p><p>Le médecin évalue si ce dispositif d’accompagnement psychologique est pour vous, ou s’il est préférable de vous orienter directement vers des soins plus spécialisés.</p>",
+    question: "Suis-je concerné(e) par le dispositif ?",
+  },
+  {
+    answer:
+      "<p>Si vous êtes inquiet(e), consultez rapidement votre médecin. Il pourra vous dire si vos troubles relèvent d’un suivi avec un psychologue ou s’il faut plutôt consulter un psychiatre ou un service spécialisé.</p><p>Si vous craignez un geste suicidaire ou de violence, appelez le SAMU ou le 3114, numéro national de prévention du suicide. Des psychologues et des infirmiers formés à la prévention du suicide sont à votre écoute 24h/24 et 7j/7.</p>",
+    question: "Que faire si je suis inquiet(e) ?",
+  },
+];