+++ src/services/faq/patient/reimbursment.ts
@@ -1,0 +1,42 @@
+/* eslint-disable max-len */
+export default [
+  {
+    answer:
+      "<p>Dans le cadre du dispositif MonPsy, et après une orientation par médecin, l’Assurance Maladie prend en charge 1 entretien d’évaluation et jusqu’à 7 séances de suivi. Le nombre de séances réalisées est adapté aux besoins du patient par le psychologue. L’accompagnement psychologique proposé est, en effet, conçu comme une intervention brève dont la littérature a montré l’efficacité et la pertinence pour les troubles psychiques légers à modérés. Il ne s’agit pas d’une psychothérapie structurée pour des troubles sévères. </p><p>A noter, chaque séance doit être réalisée à une date distincte de la précédente. L’assurance maladie ne rembourse pas si plusieurs séances sont effectuées à une même date.</p><p>Au total, vous ne pouvez bénéficier du remboursement de 8 séances au maximum par année civile. Les éventuelles séances supplémentaires ne sont pas prises en charge par l’Assurance Maladie.</p>",
+    question: "Combien de séances sont remboursées par l’Assurance Maladie ?",
+  },
+  {
+    answer:
+      "<p>Dans le cadre de ce dispositif, vous (patient) payez directement le psychologue après chaque séance ou à la fin de plusieurs séances (selon le choix du psychologue) :<ul><li><b>40€ pour la première séance</b>, l’entretien d’évaluation,</li><li><b>30€ pour chacune des séances de suivi</b> (7 au maximum).</li></ul></p><p>Le psychologue ne peut pas appliquer de dépassement d’honoraires dans le cadre de ce dispositif.</p><p>Comme lorsque vous consultez d’autres professionnels de santé, le psychologue complète et vous remet, après chaque séance ou à la fin de plusieurs séances (selon le choix du psychologue), une feuille de soins que vous envoyez à votre organisme d’assurance maladie, pour vous faire rembourser (cf. prochaine question sur le remboursement).</p><p><b>Vous bénéficiez du tiers payant obligatoire et vous n’avez pas à avancer les frais</b>, si vous êtes dans l’une des situations suivantes :<ul><li>Bénéficiaire de la Complémentaire Santé Solidaire (CSS) </li><li>Bénéficiaire de l’Aide Médicale d’Etat (AME)</li><li>Soins en lien avec une maladie : Affection de Longue Durée (ALD) ou accident causé par un tiers</li><li>Soins en lien avec une maternité (à partir du 6ème mois de grossesse)</li><li>Soins en lien avec un accident du travail ou une maladie professionnelle (AT-MP)</li></ul></p><p>Dans ces cas, le psychologue étant rémunéré directement par votre organisme d’assurance maladie, il faudra obligatoirement que :<ul><li>Vous transmettiez au psychologue qui vous suit le courrier d’adressage de votre médecin</li><li>Vous signiez la feuille de soins, même si c’est le psychologue qui la transmet à votre organisme d’assurance maladie par la suite.</li></ul></p>",
+    question: "Dois-je payer directement le psychologue ?",
+  },
+  {
+    answer:
+      "<p>Dans le cadre de ce dispositif, c’est vous (patient), sauf cas indiqués ci-dessous, qui payez directement le psychologue après chaque séance ou à la fin de plusieurs séances (selon le choix du psychologue). </p><p>En pratique :</p><p>1-Le psychologue complète et vous transmet, après chaque séance ou à la fin de plusieurs séances (selon votre choix), la feuille de soins avec les soins payés. </p><p>2- Afin d’être remboursé(e), vous transmettez à votre organisme d’assurance maladie, la feuille de soins accompagnée du courrier d’adressage de votre médecin. Lors du premier envoi de la feuille de soins correspondant à l’entretien d’évaluation, vous devez obligatoirement joindre le courrier d’adressage du médecin.</p><p>3- Vous êtes ensuite remboursé(e) par votre organisme d’assurance maladie obligatoire (60% du tarif/séance) et le cas échéant par votre complémentaire / mutuelle.</p><p>Dans le cadre du dispositif MonPsy, le psychologue vous facture les séances aux tarifs de 40€ pour l’entretien d’évaluation (= la première séance) et 30€ pour les séances de suivi. Aucun dépassement ne sera possible.</p><p>A noter, dans les prochaines années, sera développée progressivement l’utilisation de la carte vitale par les psychologues partenaires, afin de simplifier vos démarches. </p>",
+    question: "Comment me faire rembourser par l’Assurance Maladie ?",
+  },
+  {
+    answer:
+      "<p>Dans le cadre de ce dispositif, l’Assurance Maladie prend en charge 60% du coût des séances, c’est-à-dire 24€ pour la première séance et 18€ pour les séances de suivi. </p><p>Votre mutuelle ou votre assurance complémentaire, dans le cadre des contrats de complémentaire santé responsable (couvrant 95% des assurés), finance à hauteur de 40% le coût des séances.</p>",
+    question:
+      "Comment le coût des séances est-il pris en charge pour la part « complémentaire » ?",
+  },
+  {
+    answer:
+      "<p>Si vous êtes dans une des situations suivantes, vous bénéficiez du tiers ayant obligatoire et vous ne faites donc pas l’avance de frais : <ul><li>Bénéficiaire de la Complémentaire Santé Solidaire (CSS) </li><li>Bénéficiaire de l’Aide Médicale d’Etat (AME)</li><li>Soins en lien avec une maladie : Affection de Longue Durée (ALD) ou accident causé par un tiers</li><li>Soins en lien avec une maternité (à partir du 6ème mois de grossesse)</li><li>Soins en lien avec un accident du travail ou une maladie professionnelle (AT-MP)</li></ul></p><p>En pratique :</p><p>1-Votre médecin précise sur le courrier d’adressage que vous êtes dans une des trois dernières situations (des soins en lien avec une maladie, une maternité ou un AT-MP). </p><p>2-Vous apportez, lors de la première séance chez le psychologue, votre attestation de droits à jour (= attestation de carte Vitale papier indiquant vos droits) et vous donnez au psychologue le courrier d’adressage de votre médecin.</p><p>3-A la fin de chaque séance, vous signez la feuille de soins (que vous laissez au psychologue) et <b>vous ne réglez rien à la fin de la séance</b>. Le psychologue est rémunéré directement par votre organisme d’assurance maladie. </p>",
+    question:
+      "Dans quels cas, suis-je exonéré(e) d’avance de frais, et je ne paie donc rien à la fin des séances ?",
+  },
+  {
+    answer:
+      "<p>Dans le cadre du dispositif MonPsy, le psychologue doit vous facturer les séances aux tarifs conventionnés et remboursés par l’Assurance Maladie : 40€ pour la première séance (l’entretien d’évaluation) et 30€ pour les séances suivantes (les séances de suivi). Aucun dépassement n’est possible.</p><p>Si le psychologue applique des tarifs non conventionnés (dépassements d’honoraires), la feuille de soins est rejetée automatiquement par les organismes payeurs (Assurance Maladie et complémentaires santé) et vous est retournée. Pour être pris en charge, le psychologue devra refaire une feuille de soins avec le bon tarif et vous rembourser la différence.</p>",
+    question:
+      "Le psychologue peut-il pratiquer des dépassements d’honoraires (c’est-à-dire me demander de payer une séance plus chère que le tarif pris en charge par l’Assurance Maladie) ?",
+  },
+  {
+    answer:
+      "<p>Si votre médecin estime que le recours au psychologue est en lien avec une de vos ALD et que vous présentez une déficience ou incapacité justifiant la prescription d’un transport sanitaire, alors les transports peuvent être pris en charge dans les conditions habituelles (article R.322-10 du code de la sécurité sociale). C’est le seul cas de prise en charge.</p>",
+    question:
+      "Puis-je bénéficier de la prise en charge des transports pour me rendre chez un psychologue ?",
+  },
+];