+++ src/services/faq/psy/eligibility.ts
@@ -1,0 +1,32 @@
+/* eslint-disable max-len */
+export default [
+  {
+    answer:
+      "<p>Pour devenir psychologue partenaire du dispositif MonPsy, il est nécessaire de : <ul><li>Etre inscrit sur le registre ADELI attestant du titre de psychologue;</li><li>Disposer d’une expérience professionnelle en psychologie clinique ou en psychopathologie de trois ans minimum ;</li><li>Avoir une expérience ou un diplôme permettant d’attester d’un parcours consolidé en psychologie clinique ou en psychopathologie.</li></ul></p><p>Ces critères ont été définis en accord avec les représentants des psychologues.</p>",
+    question: "Quelles sont les conditions de participation au dispositif ?",
+  },
+  {
+    answer:
+      "<p>Le dispositif ne concerne que les professionnels disposant du titre règlementé de psychologue. Certains psychologues détiennent aussi d’autres qualifications qu’ils peuvent mentionner dans leur formulaire de candidature.</p><p>Les psychothérapeutes agréés par les ARS ou les psychanalystes, s’ils n’ont pas un parcours attesté/ consolidé en psychologie clinique ou en psychopathologie et ne sont pas inscrits au registre ADELI, ne sont pas éligibles au dispositif MonPsy.</p>",
+    question:
+      "Faut-il nécessairement être psychologue ? Qu’en est-il des psychothérapeutes agréés par les ARS ?",
+  },
+  {
+    answer:
+      "<p>Le numéro ADELI est obligatoire, car il permet d’attester que son détenteur est autorisé à faire usage du titre de psychologue. Le numéro SIRET n’offre pas une telle garantie. Les démarches pour obtenir le numéro ADELI doivent être réalisées auprès de l’ARS de son lieu d’exercice.</p><p>Le numéro ADELI commence par le numéro de département de votre exercice principal. Si tel n’est pas le cas, rapprochez-vous de votre ARS pour obtenir un numéro ADELI correspondant.</p>",
+    question:
+      "Je ne dispose pas de numéro Adeli mais d’un numéro Siret. Puis-je participer au dispositif ?",
+  },
+  {
+    answer:
+      "<p>Si votre employeur est d’accord, il est tout à fait possible de participer au dispositif en tant que psychologue salarié(e) d’une structure d’exercice coordonné (maison de santé pluriprofessionnelle ou centre de santé). Il est alors nécessaire de l’indiquer dans le questionnaire de candidature et la convention sera tripartite (psychologue, structure et Assurance Maladie).</p><p>Il est aussi possible de devenir un psychologue partenaire dans le cadre de son activité libérale et de son activité salariée.</p>",
+    question:
+      "Puis-je participer au dispositif en tant que salarié(e) d’une structure d’exercice coordonné ?",
+  },
+  {
+    answer:
+      "<p>Si vous souhaitez être référencé(e) pour accompagner les enfants ou les adolescents, vous devez avoir une expérience clinique auprès d’enfants ou d’adolescents. Il vous revient d’apprécier si cette expérience est suffisante pour prendre en charge des enfants et des adolescents. En tant que psychologue partenaire, votre souhait de pouvoir recevoir des enfants et adolescents sera indiqué dans l’annuaire accessible au public.</p>",
+    question:
+      "Puis-je participer au dispositif si je n’ai pas d’expérience avec les enfants et/ou les adolescents ?",
+  },
+];